requirements.txt
mysql_connector_repackaged==0.3.1
PySide6==6.5.2
PySide6_Addons==6.5.2
PySide6_Essentials==6.5.2
webdriver-manager==4.0.1